requirements.txt
discord.py
uvicorn